+++ app/graphql/application_schema.rb
@@ -1,17 +1,17 @@
 require "graphql/batch"
 
 class ApplicationSchema < GraphQL::Schema
-  mutation(Types::MutationType)
-  query(Types::QueryType)
-
-  use GraphQL::Batch
-  use GraphQL::PersistedQueries, store: :redis_with_local_cache, compiled_queries: true
+  mutation Types::MutationType
+  query Types::QueryType
 
   max_complexity 100
   max_depth 10
 
+  use GraphQL::Batch
+  use GraphQL::PersistedQueries, store: :redis_with_local_cache, compiled_queries: true
+
   rescue_from(ActiveRecord::RecordNotFound) do |_err, _obj, _args, _ctx, field|
     raise GraphQL::ExecutionError.new("#{field.type.unwrap.graphql_name} not found",
-                                      extensions: { message: "Not Found", status: 404, code: :not_found })
+      extensions: { message: "Not Found", status: 404, code: :not_found })
   end
 end
