--- app/graphql/resolvers/public/activities.rb
@@ -1,15 +1,0 @@
-module Resolvers
-  module Public
-    class Activities < Resolvers::Base
-      include SkipAuthentication
-
-      description "Fetch public activities"
-
-      type Types::Public::ActivityType.connection_type, null: true
-
-      def fetch_data
-        Activity.public_events
-      end
-    end
-  end
-end
